rename cache_status_jel to cache_status and use it. Fix #6411 (#6878)
